Appending content to an array instead of a string
seems to speed-up rendering of a small template by 3000x.

Looks to good to be true. Am I missing something? :)

```
Warming up --------------------------------------
              string     2.964k i/100ms
               array     2.955k i/100ms
Calculating -------------------------------------
              string     30.570k (±11.0%) i/s -    151.164k in
5.006935s
               array    103.274M (±22.5%) i/s -    422.819M in
4.941287s

Comparison:
               array: 103273902.7 i/s
              string:    30569.9 i/s - 3378.29x  slower
```