@jeremyevans 2017-02-01
In general, template compilation speed is not all that important, the execution speed of the compiled template is what matters most.

You could just do this if you wanted to use an array:

```ruby
Erubi::Engine.new(template, :src=>[]).src.join.freeze
```

That should work without any changes.

The benchmark results do look to good to be true, that's because they are.  Remove the block argument from the `x.report("array") do`  line, then you get more reasonable output:

```
Warming up --------------------------------------
              string     1.203k i/100ms
               array     1.165k i/100ms
Calculating -------------------------------------
              string     12.411k (_ 0.6%) i/s -     62.556k in   5.040579s
               array     12.100k (_ 0.6%) i/s -     60.580k in   5.006985s

Comparison:
              string:    12410.9 i/s
               array:    12099.6 i/s - 1.03x  slower
```

If you are using MRI and not JRuby, and the warmup times are significantly different than the calculation times, you should suspect something may be wrong.